Fixes policies set-version
