+++ scripts/plugin-pack.js
@@ -1,7 +1,7 @@
 const {existsSync} = require(`fs`);
 
-if (existsSync(`${__dirname}/../packages/plugin-pack/bin/@berry/plugin-pack.js`)) {
-  module.exports = require(`${__dirname}/../packages/plugin-pack/bin/@berry/plugin-pack.js`);
+if (existsSync(`${__dirname}/../packages/plugin-pack/bundles/@berry/plugin-pack.js`)) {
+  module.exports = require(`${__dirname}/../packages/plugin-pack/bundles/@berry/plugin-pack.js`);
 } else {
-  module.exports = require(`${__dirname}/local/berry-plugin-pack.js`);
+  module.exports = require(`${__dirname}/../packages/plugin-pack/bin/@berry/plugin-pack.js`);
 }
