+++ scripts/run-yarn.js
@@ -1,7 +1,7 @@
 const {existsSync} = require(`fs`);
 
-if (existsSync(`${__dirname}/../packages/berry-cli/bin/berry.js`)) {
-  require(`${__dirname}/../packages/berry-cli/bin/berry.js`);
+if (existsSync(`${__dirname}/../packages/berry-cli/bundles/berry.js`)) {
+  require(`${__dirname}/../packages/berry-cli/bundles/berry.js`);
 } else {
-  require(`${__dirname}/local/berry.js`);
+  require(`${__dirname}/../packages/berry-cli/bin/berry.js`);
 }
